chore: remove old files and migrate story
